TX queue checks removed from ASF code
ASF is doing the checks for Tx queue based on the Rx queue the packet was reveived on
but with the RX_TX_BD_XNGE patch or the SW packet steering patch the Tx queue is selected
based on the processor-id on which the packet is processed. Making the Tx queue checks
irrelevent in the ASF.
For cases where RX_TX_BD_XNGE or SW packet steeting patch is not used, configuring the
queue mapping as 0.
BugId : ENGR140280

Signed-off-by: Rajan Gupta <rajan.gupta@freescale.com>