asf_timer: Patch to take care of race condition for timer node deletion.
This patch fixes the issue arising out of race condition. The race
condition was coming in one case where timer stop request being
triggered at core0 and the timer belongs to core1. In that case
the timer was added to the reclaim queue for core0 wherein. After
stopping the timer the calling module cleans up its data structures.
When the timer fires on core1 it used to get error in callback and
used to schedule the RCU. Later in the code it cleans up the reclaim
queue which does memset of the timer node. Now when RCU gets called
the function gets initialized to NULL and hence the system used to
crash. Now the call_rcu is being done only in case if the timer is
oneshot timer.

Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>