Fix for packet flowing through Linux when firewall is deleted.
Problem Summary: Packets are flowing through Linux i.e. flows are not getting offloaded when Firewall rule is deleted
before older session is timed out in Linux.
Solution:
When a firewall rule is set, instead of deleting the flow from ASF its drop bit is set.
Flow is deleted when it is timed out. When firewall rule is deleted, flow drop bit is reset.
This new approach also solves the problem of packet drop in both direction while firewall rule is set for one direction.
Signed-off-by: Nikhil Agarwal <nikhil.agarwal@freescale.com>