asf: T4240: Frame TX mechanism is synced with current DPAA Eth driver.
In T4240QDS, DPA maintains a new frame queue which is used for the recycled
buffers only. As per the implementation, when a frame is to be transmitted
then frame must be placed in this newly added frame queue so that after
transmission, same buffer can be used again. In Older implementation frame
was being written to other frame queues that's why same can not be used further.
So as per the current DPA eth driver same changes are required in ASF also.

CR:ENGR00274481

Signed-off-by: Sunil Kumar Kori <Sunil.Kori@freescale.com>
Change-Id: I1ab5caeb555d3ad334194a5dbe75ebdda0670822
Reviewed-on: http://git.am.freescale.net:8181/4110
Reviewed-by: Gupta Rajan-B15745 <rajan.gupta@freescale.com>
Tested-by: Gupta Rajan-B15745 <rajan.gupta@freescale.com>