AH-SHA1 policy: Decapsulation was not occurring in case of IPv4-in-IPv6 and IPv6-in-IPv4.
On sending IPv4 traffic through IPv6 tunnel and IPv6
packet through IPv4 tunnel, decapsulation was not occuring.
Reason was that Outer header was being decided on the basis
of next header of the encapsulated packet.

It was resolved by checking outer IP header version whether
it is 4 or 6.

Signed-off-by: Ganga Negi <ganga.negi@freescale.com>
Change-Id: Id53c3425a96218f5baca32bf7ad8ba760936847e
CR: ENGR00313834
Reviewed-on: http://git.am.freescale.net:8181/12429
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>