ASF: setting gianfar_destructor to NULL before sending packet to linux.
After copying and sending the packet to linux, linux was calling
our 'gianfar_destructor' pointer function (which was set for ASF leg).
This should be set to NULL before sending the packet to linux.
There was a memory leak due to this. Also function to free the
skb should be 'kfree_skb' to free the skb allocated for linux in
function 'asfctrl_ipsec_fn_VerifySPD'.

Change-Id: I7b3743079ccc0491db0d359a5e082ea3f1f41e9f
Signed-off-by: Nipun Gupta <nipun.gupta@freescale.com>
Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/17122
Reviewed-by: Hemant Agrawal <hemant@freescale.com>