asf_ipsec: Patch to fix the crash seen in case of AH IPsec.
In case of dynamic re-keying the SA gets deleted after
hard expiry of timer. In case few packets are still with
CAAM upon retruning context from CAAM the check for SA
validity was missing. So due to this the system was crashing
in case the SA gets deleted. This patch checks for NULL SA and
in case it is NULL the packet is dropped.

CR:ENGR276702
CR:ENGR270320

Change-Id: I68666e7867b80eb07fc01518ffb8da518d62338e
Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/17142
Reviewed-by: Hemant Agrawal <hemant@freescale.com>