asf_ipsec_ah: Patch to fix the issue of ping6 not working.
This patch fixes the issue of ping6 not working when the
AH tunnel is configured with IPv6 in IPv4 mode. The reason
for ping not working was that the outer header was removed
based on the inner header information. The code has been
updated to take care of this and now it is updating the
skb->data based on outer ip header information.

CQ:ENGR00286712

Change-Id: Ied01a85cba970805d60021f0532d225d79cc2c44
Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/17145
Reviewed-by: Hemant Agrawal <hemant@freescale.com>