ASFIPSEC: Crash on making interface down.
In case of ESP/AH IPv4-in-IPv4/IPv6-in-IPv6 on doing egress interface
down we get crash.

After making egress interface of the DUT down, IPSec L2Blob details were
not getting refreshed before sending it to hard xmit function. Hence, we
were getting crash there on accessing net device pointer. Resolved by
refereshing IPSec L2Blob if dev is down and in case of AH putting a check
for network device whether it is NULL or not.

Signed-off-by: Ganga Negi <ganga.negi@freescale.com>
CR:ENGR00329874
Change-Id: Ifa063230ebbdc7bc7073368ec3ef3490f90c2c2c
Reviewed-on: http://git.am.freescale.net:8181/18202
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22641