ASFIPSEC: Crash on making interface down.
In case of ESP/AH IPv4-in-IPv4/IPv6-in-IPv6 on doing egress interface
down we get crash.

After making egress interface of the DUT down, IPSec L2Blob details were
not getting refreshed before sending it to hard xmit function. Hence, we
were getting crash there on accessing net device pointer. Resolved by
refereshing IPSec L2Blob if dev is down and in case of AH putting a check
for network device whether it is NULL or not.

Signed-off-by: Ganga Negi <ganga.negi@freescale.com>
Change-Id: I57453fc1a4691d71899333a78fcfaed3b4bb3031
CR:ENGR00329874
Reviewed-on: http://git.am.freescale.net:8181/18202
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22641
Reviewed-on: http://git.am.freescale.net:8181/29941