ASF_IPSEC: Fix for SA leak(Orphaned SA wothout Container)
During the deletion of container corresponding data structures are
not cleared properly from SA. This is fixed by calling corresponding
UnMapSA function before deleting container,if SA is mapped to more
than one policy. If SA is mapped to single policy, delete that
SA in the ASF before deleting the container.

This patch also removes the extra definiations of some functions.

CR:ENGR00312378

Change-Id: I5fa9064026178524ebca80151bced5b2f80724d1
Signed-off-by: Sridhar Pothuganti <sridhar.pothuganti@freescale.com>
Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/17783