ASFIPSEC: Fix for crash seen on T4240 in case of ping
On T4240, crash was coming because in secf_outAHComplete
the skb->dev is coming NULL in case of ping
packet(self generated) and when we try to find the dpa
pool based on this dev, then crash occurs.
Correct check is introduced to get the dpa pool only when
it is required.

Change-Id: I159563d090d903f77ed7ab1ea0dd29c7ccf221ce
Signed-off-by: Sahil Malhotra <sahilmalhotra@freescale.com>
CR:ENGR00331394
Reviewed-on: http://git.am.freescale.net:8181/19017
Reviewed-by: Arun Pathak <arun.pathak@freescale.com>
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>