ASFIPSEC: struct addr in scatter_gather_entry_t is changed.
Current code uses the shift operation to store the physical
addresses.
e.g.
addr = (sgt.addr_hi)<<32 + sgt.addr_lo;

addr_hi is uint_32, the value at the lowest bit position will
be lost if it is left shift it by 32 bits.

With this patch the bit-wise operators are not needed.

Change-Id: I7f0fa7803fbcb977b7de5f46950b06c27f82ea6b
Signed-off-by: Sahil Malhotra <sahilmalhotra@freescale.com>
CR:ENGR00333204
Reviewed-on: http://git.am.freescale.net:8181/19956
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>