asf: Patch to fix tail pointer-offset issue in 64-bit machine.
skb->tail is assigned a pointer to tail which is
valid for 32-bit machine but in case of 64-bit
machine it is offset not the pointer to tail.
So in this patch tail and end pointer is set
using wrapper function of linux which will take
care of the 32-bit/64-bit machine.

CR:ENGR00331377

Signed-off-by: Alok Makhariya <B46187@freescale.com>
Change-Id: I3a47f32fbfa997873d69ec52709152e1e9c2907d
Reviewed-on: http://git.am.freescale.net:8181/20139
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>