ASFIPSEC: TOS is not correctly copied in IPv4 in IPv6 case
In case of SEC4x and PROTO_OFFLOAD, when traffic for IPv4 in
IPv6 is sent then TOS field in IPv4 packet is not getting
copied in the Outer IPv6 packet.
This is because when packet comes out after encryption,
ip6 header flow_label field contains some garbage value and
these fields were not cleared before updating flow_label.
In this patch the first 4 bits of flow_label is first
cleared and then updated.

Change-Id: Ib5faa76f0317ffa144e7e82a00f65e2e1b96fd08
Signed-off-by: Sahil Malhotra <sahilmalhotra@freescale.com>
CR:ENGR00334611
Reviewed-on: http://git.am.freescale.net:8181/20983
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>