Applying Firewall rule in running traffic gives CRASH.
On configuring firewall in running traffic due to
accessing NULL pointer skb_rtable(skb) we got crash.
Function ip_route_input() could not find route with
'lo' device passed to it, because rp filter was ON,
due to which skb_rtable(skb) was NULL.

On passing ingress device pointer to ip_route_input()
for lookup in routing table fixed this issue.

Signed-off-by: Ganga Negi <ganga.negi@freescale.com>
Change-Id: I900fdc0eae58e1cc80bf280dadde3b3d2495ba19
CR:ENGR00324697
Reviewed-on: http://git.am.freescale.net:8181/29943
Reviewed-by: Sandeep Malik <Sandeep.Malik@freescale.com>
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>